% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreg_blau.R
\name{aggreg_blau}
\alias{aggreg_blau}
\title{Compute a Blau index}
\usage{
aggreg_blau(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector. Set of values (0 or 1) for which the Blau index should be computed.}

\item{na.rm}{Logical. Should NAs be omitted.}
}
\value{
A Blau index for the numeric vector
}
\description{
Computes the Blau index.
}
\author{
Nicolas Mangin
}
