% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreg_cvar.R
\name{aggreg_cvar}
\alias{aggreg_cvar}
\title{Compute a coefficient of variation}
\usage{
aggreg_cvar(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector. Set of values for which the coefficient of variation should be computed.}

\item{na.rm}{Logical. Should NAs be omitted.}
}
\value{
A coefficient of variation for the numeric vector
}
\description{
Computes the coefficient of variation.
}
\author{
Nicolas Mangin
}
