% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim.R
\name{aim}
\alias{aim}
\title{Aim}
\usage{
aim()
}
\value{
Character string echoing the terminal.
}
\description{
Sbalbalbaab.
}
\examples{
\dontrun{
aim()
}
}
\author{
Jan Netik \cr

Department of Psychology, \cr Faculty of Arts, \cr
Charles University, \cr Czech Republic \cr

\email{netikja@gmail.com} \cr
}
