% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_tm_quality.R
\name{assess_tm_quality}
\alias{assess_tm_quality}
\title{Return various metrics to evaluate topic model performance for different topic numbers.}
\usage{
assess_tm_quality(doc_to_term = NA, topic_nbr = c(0),
  prevalence = NA, init.type = "Spectral", seed_nbr = 1,
  keywords = NA, topic_per_doc = 1, terms_per_topic = 5)
}
\arguments{
\item{doc_to_term}{dtm object.}

\item{topic_nbr}{Numeric vector}

\item{prevalence}{Formula}

\item{init.type}{Character.}

\item{seed_nbr}{Numeric vector}

\item{keywords}{Tibble.}

\item{topic_per_doc}{Integer.}

\item{terms_per_topic}{Integer.}
}
\value{
A tibble with various metrics of tm quality for each possible combination of the topic number and seed
}
\description{
Return various metrics to evaluate topic model performance for different topic numbers.
}
