% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostr.R
\name{boostr}
\alias{boostr}
\title{Parallelize a function applied over the rows of a tibble.}
\usage{
boostr(x, FUN, intovar, threads, ...)
}
\arguments{
\item{x}{Tibble. Data on which rows the function should be applied.}

\item{FUN}{Function. Function to apply on the data/}

\item{intovar}{Character. Variable where the results should be stored.}

\item{threads}{Integer. Number of parallel threads.}

\item{...}{Additional arguments to be passed.}
}
\value{
The initial tibble with the results in the variable specified with intovar.
}
\description{
Parallelize a function applied over the rows of a tibble.
}
