% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bos_dictionaries.R
\docType{data}
\name{bos_dictionaries}
\alias{bos_dictionaries}
\title{Dictionaries for lexical analyses}
\format{A tibble}
\source{
\url{http://www.lexiconista.com/datasets/lemmatization/}
}
\usage{
data("bos_dictionaries")
}
\description{
A tibble with patterns classified in various categories.
}
\keyword{LIWC}
\keyword{dictionary,}
