% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bos_en_lemmas.R
\docType{data}
\name{bos_en_lemmas}
\alias{bos_en_lemmas}
\title{English lemmatization}
\format{A tiblle with 66,792 observations and 2 variables ("word" and "term")}
\source{
\url{http://www.lexiconista.com/datasets/lemmatization/}
}
\usage{
data("bos_en_lemmas")
}
\description{
A table matching words with lemmas.
}
\keyword{lemmatization,}
\keyword{stem}
