% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bos_format_ascii.R
\name{bos_format_ascii}
\alias{bos_format_ascii}
\title{Remove html tags, remove double spaces, convert to ASCII.}
\usage{
bos_format_ascii(text)
}
\arguments{
\item{text}{character. String to be cleaned.}
}
\value{
A clean string.
}
\description{
Remove html tags, remove double spaces, convert to ASCII.
}
