% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bos_sentiments.R
\docType{data}
\name{bos_sentiments}
\alias{bos_sentiments}
\title{Sentiments}
\format{A tibble}
\source{
\url{http://www.lexiconista.com/datasets/lemmatization/}
}
\usage{
data("bos_sentiments")
}
\description{
A tibble with patterns classified in various categories.
}
\keyword{LIWC}
\keyword{dictionary,}
