% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bos_symbols.R
\docType{data}
\name{bos_symbols}
\alias{bos_symbols}
\title{List of symbols to be removed}
\format{A character string with symbols.}
\usage{
data("bos_symbols")
}
\description{
List of symbols to be removed
}
\keyword{cleaning}
\keyword{text}
