% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bos_toascii.R
\docType{data}
\name{bos_toascii}
\alias{bos_toascii}
\title{Map non ascii to ascii}
\format{A tiblle with 66,792 observations and 2 variables ("word" and "term")}
\source{
\url{http://www.lexiconista.com/datasets/lemmatization/}
}
\usage{
data("bos_toascii")
}
\description{
A table matching characters with their ascii counterpart
}
\keyword{lemmatization,}
\keyword{stem}
