% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bow_from_doclist.R
\name{bow_from_doclist}
\alias{bow_from_doclist}
\title{Transform the documents stored in a dataframe in a bag of words.}
\usage{
bow_from_doclist(x, rm_words = NULL, min_nchar = 3,
  match_term = NULL)
}
\arguments{
\item{x}{character. Content to transform in a bag of words.}

\item{rm_words}{character vector. List of words (or n-tokens) to remove.}

\item{min_nchar}{integer. Number of characters below which the word is removed.}

\item{match_term}{dataframe. Table with two variables: "word" for the string in the content and "term" for its categorization. Match words to force desired completions.}
}
\value{
A tibble with the id of the document, words, stems, lemmas, terms, counts, and proportions.
}
\description{
Transform the documents stored in a dataframe in a bag of words.
}
