% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_friendly.R
\name{build_binaries}
\alias{build_binaries}
\title{Build a set of binaries, waiting until complete}
\usage{
build_binaries(filenames, host, port = 8765, ..., dest = tempfile())
}
\arguments{
\item{filenames}{Vector of filenames of the source packages}

\item{host}{A hostname for the build server, or \code{FALSE} to
build locally.}

\item{port}{Port number for the build server}

\item{...}{Arguments passed through to the \code{build} method of
\code{\link{buildr_client}}; includes \code{dest} (place to put
binaries), \code{poll} (frequency to poll for job completion),
\code{timeout} (time until giving up), \code{verbose} (controls
printed output) and \code{log_on_failure} (print logs if a build
fails).}
}
\value{
A character vector of filenames
}
\description{
Build a set of binaries, waiting until complete
}

