% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildr.R
\name{build_package}
\alias{build_package}
\title{Internal build function}
\usage{
build_package(file, root)
}
\arguments{
\item{file, root}{arguments}
}
\description{
Internal build function
}
\keyword{internal}

