% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildr-package.R
\docType{package}
\name{buildr-package}
\alias{buildr}
\alias{buildr-package}
\title{buildr: Organize & Run Build Scripts Comfortably}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Working with reproducible reports or any other
    similar projects often requires to run the script that builds the
    output file in a specified way. One can become tired from repeatedly
    switching to the build script and sourcing it. The 'buildr' package
    does this one simple thing via 'RStudio' addin – user can set up the
    keyboard shortcut and run the build script with one keystroke anywhere
    anytime. The second way is to pass buildr() command to console which
    does the same thing. Both ways source the build.R (case insensitive)
    file present in the current working directory.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://netique.github.io/buildr}
  \item Report bugs at \url{https://github.com/netique/buildr/issues}
}

}
\author{
\strong{Maintainer}: Jan Netik \email{netikja@gmail.com} (\href{https://orcid.org/0000-0002-3888-3203}{ORCID})

}
\keyword{internal}
