% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildr.R
\name{buildr}
\alias{buildr}
\title{Create buildr object}
\usage{
buildr(path, n_workers = 0L, quiet_workers = FALSE)
}
\arguments{
\item{path}{Path to store all our files in.  This could get
reasonably large over time.}

\item{n_workers}{Number of workers to spawn.  This can be zero, in
which case you will have to spawn workers yourself.}

\item{quiet_workers}{Make the workers quiet?  If FALSE, the
workers will log to the directory \code{file.path(path,
"worker_log")} with one file per worker.  If \code{TRUE} all
worker output will be interleaved in standard output on this
terminal (which might get confusing/annoying).}
}
\description{
Create the buildr workhorse function.  This does most of the hard
work of running the queue, submitting jobs and the like, but it
does not run a loop (see \code{buildr_server} for that).
}

