% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{buildr_client}
\alias{buildr_available}
\alias{buildr_client}
\title{buildr Client}
\usage{
buildr_client(host, port = 8765)

buildr_available(host, port = 8765)
}
\arguments{
\item{host}{Hostname that the buildr server is running on}

\item{port}{Port that the buildr server is running on (the default
here matches the buildr server default)}
}
\description{
Client for submitting and retrieving packages from a buildr
server.
}

