% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{buildr_server}
\alias{buildr_server}
\title{Run buildr server}
\usage{
buildr_server(path, n_workers, quiet_workers = FALSE, host = "0.0.0.0",
  port = 8765, loop = TRUE)
}
\arguments{
\item{path}{Path to store all our files in.  This could get
reasonably large over time.}

\item{n_workers}{Number of workers to spawn.  This can be zero, in
which case you will have to spawn workers yourself.}

\item{quiet_workers}{Make the workers quiet?  If FALSE, the
workers will log to the directory \code{file.path(path,
"worker_log")} with one file per worker.  If \code{TRUE} all
worker output will be interleaved in standard output on this
terminal (which might get confusing/annoying).}

\item{host}{A string that is a valid IPv4 address that is owned by
this server, or \code{"0.0.0.0"} to listen on all IP addresses.}

\item{port}{The port to listen on.}

\item{loop}{Run in a loop?  This is generally the right move
unless you plan on handling requests (with
\code{httpuv::service} manually).}
}
\description{
Run a buildr server.  This runs a http server that will listen for
requests to build packages.
}

