% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_string.R
\name{clean_string}
\alias{clean_string}
\title{Simplify and clean a text.}
\usage{
clean_string(x)
}
\arguments{
\item{x}{character. Text to be cleaned.}

\item{language}{character string. Language for the stopwords to remove.}
}
\value{
A clean text.
}
\description{
Simplify and clean a text.
}
