% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_in_window.R
\name{count_in_window}
\alias{count_in_window}
\title{Count occurrences of a focal regex expression (near an adjacent one if specified) in the result of a udpipe part-of-speech tagging process.}
\usage{
count_in_window(text, focal, dictionary, window = c(3, 3))
}
\arguments{
\item{text}{character. Text.}

\item{focal}{regex pattern. Pattern around which the adjacent pattern has to be detected within the window.}

\item{dictionary}{dataframe. 2 variables: pattern and category. Specify and classify regex patterns.}

\item{window}{numeric vector. Two values: one indicating the beginning of the window from the focal pattern, and one indicating the end.}
}
\value{
A numeric value indicating how many times the focal pattern occurred.
}
\description{
Count occurrences of a focal regex expression (near an adjacent one if specified) in the result of a udpipe part-of-speech tagging process.
}
