% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_makefile.R
\name{edit_makefile}
\alias{edit_makefile}
\title{Edit Makefile}
\usage{
edit_makefile()
}
\value{
No return value. Called for side effect.
}
\description{
Opens \code{Makefile}, if present in the project root.
}
\seealso{
The \href{https://www.gnu.org/software/make/manual/html_node/}{documentation for GNU Make}.
}
