% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_shortcuts.R
\name{edit_shortcuts}
\alias{edit_shortcuts}
\title{Show RStudio Keyboard Shortcuts Popup}
\usage{
edit_shortcuts()
}
\value{
No return value. Called for side effect.
}
\description{
Shows popup window with RStudio keyboard shortcuts. Uses \code{rstudioapi}.
Applicable only in RStudio and in interactive session.
}
\details{
You can quicky reach out solicited addin function by typing it in the
\code{Filter...} box in the very top of the popup window. Then double click at the
blank space just next to the addin function name and press down desired key
or key combination. Apply the changes and from now on, just call the function
with one keystroke.
}
\examples{
\dontrun{
edit_schortcuts()
}
}
