% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map all recorded accidents of a specified year on a map graphic of a specified state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric value corresponding to the state number}

\item{year}{A numeric value corresponding to the csv file to be imported}
}
\value{
Returns a side effect map graphic of the state with all recorded
    accidents represented as points on the same map.
}
\description{
This function takes a state number and year to import and subset the corresponding csv
file. Records with missing longitudinal and latitudinal data are assigned default values.
A map graphic of the specified state is generated with points representing individual
records mapped on top.
}
\note{
If provided state number is not found in the data frame, an error is returned.
   If specified subset of data has no records associated with it, a message is returned.
}
\examples{
\dontrun{fars_map_state(1, 2013)}

}
