% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faultlines_demdiff.R
\name{faultlines_demdiff}
\alias{faultlines_demdiff}
\title{Compute a score of demographic difference using the procedure suggested by Li and Hambrick (2005) for the study of faultlines.}
\usage{
faultlines_demdiff(x, subgroup)
}
\arguments{
\item{x}{dataframe. Table informing about team members' characteristics (all team members should belong to the same team), numeric or binary. Ideally, all values should be scaled between 0 and 1 before running the analysis.}

\item{subgroup}{character. Name of the variable indicating subgroups memberships.}
}
\value{
A dataframe with one row (for the whole group) containing the index of demographic difference (DemDiff).
}
\description{
Compute a score of demographic difference using the procedure suggested by Li and Hambrick (2005) for the study of faultlines.
}
\references{
Li, J., and D. C. Hambrick. 2005. Factional groups: A new vantage on demographic faultlines, conflict, and disintegration in work teams. Academy of Management Journal 48 (5): 794–813.
}
\seealso{
make_0_to_1

find_subgroups
}
