% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faultlines_dsg.R
\name{faultlines_dsg}
\alias{faultlines_dsg}
\title{Computes the distances within and between sub-groups based on the number of common attributes (how many times members of a subgroup fall in the same category) for faultline analyses (personal procedure)}
\usage{
faultlines_dsg(x, subgroup)
}
\arguments{
\item{x}{dataframe. Binary variables indicating the categorical memberships of group members.}

\item{subgroup}{character. Name of the variable indicating subgroup membership.}
}
\value{
A dataframe with one row per attribute specifying the contribution of each attribute to subgroups internal alignements, cross group alignement, and faultline.
}
\description{
Computes the distances within and between sub-groups based on the number of common attributes (how many times members of a subgroup fall in the same category) for faultline analyses (personal procedure)
}
\seealso{
make_discrete

find_subgroups
}
