% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faultlines_fau.R
\name{faultlines_fau}
\alias{faultlines_fau}
\title{Compute distances within and between sub-groups according to the procedure designed by Thatcher, Jehn, and Zanutto (2003) for faultlines analyses.}
\usage{
faultlines_fau(x, subgroup)
}
\arguments{
\item{x}{dataframe. Table informing about team members characteristics, numeric or binary. Ideally, all values should be scaled between 0 and 1 before running the analysis.}

\item{subgroup}{character. Name of the variable indicating subgroups memberships.}
}
\value{
A dataframe specifying the distances between subgroups centroids (FAU_CGD), the distances within sub-group (FAU_SGH), the distances within the whole group (FAU_GH) the proximity between sub-groups (DSG_CGA_'subgp_base'), the faultline score (FAU_strength) and the faultline distance (FAU_dist)
}
\description{
Compute distances within and between sub-groups according to the procedure designed by Thatcher, Jehn, and Zanutto (2003) for faultlines analyses.
}
\references{
Thatcher, S. M. B., K. A. Jehn, and E. Zanutto. 2003. Cracks in Diversity Research: The Effects of Diversity Faultlines on Conflict and Performance. Group Decision and Negotiation 12 (3): 217–241.
}
\seealso{
make_0_to_1

find_subgroups
}
