% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faultlines_fls.R
\name{faultlines_fls}
\alias{faultlines_fls}
\title{Computes the distances within and between sub-groups according to the procedure designed by Shaw (2004) for faultline analyses.}
\usage{
faultlines_fls(x, subgroup)
}
\arguments{
\item{x}{dataframe. Binary variables indicating the categorical memberships of group members.}

\item{subgroup}{character. Name of the variable informing about subgroup membership.}
}
\value{
A dataframe specifying the subgroups internal alignement (FLS_IA_0, FLS_IA_1, FLS_IA), the cross-subgroups alignment (FLS_CGA), and the faultline score (FLS).
}
\description{
Computes the distances within and between sub-groups according to the procedure designed by Shaw (2004) for faultline analyses.
}
\references{
Shaw, J. B. 2004. The Development and Analysis of a Measure of Group Faultlines. Organizational Research Methods 7 (1): 66–100.
}
\seealso{
make_discrete

find_subgroups
}
