% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_replace_pattern.R
\name{files_replace_pattern}
\alias{files_replace_pattern}
\title{Replace specific strings of characters in a set of documents.}
\usage{
files_replace_pattern(path, pattern, replacement)
}
\arguments{
\item{path}{Character. Absolute path to the files.}

\item{pattern}{Character. Pattern to be replaced.}

\item{replacement}{Character. Pattern to be put instead.}
}
\value{
Replace in the same folder the documents where the string was found by the same documents where the string was replaced.
}
\description{
Replace specific strings of characters in a set of documents.
}
