% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_na.R
\name{fill_na}
\alias{fill_na}
\title{Impute numeric values with the mean and categorical values with the most frequent category.}
\usage{
fill_na(x)
}
\arguments{
\item{x}{dataframe with missing values.}
}
\value{
A dataframe without missing values.
}
\description{
Impute numeric values with the mean and categorical values with the most frequent category.
}
