% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ascii.R
\name{format_ascii}
\alias{format_ascii}
\title{Make sure all characters are ASCII}
\usage{
format_ascii(x)
}
\arguments{
\item{x}{character vector. Texts to clean.}
}
\value{
A clean character vector.
}
\description{
Make sure all characters are ASCII
}
