% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_0_to_1.R
\name{make_0_to_1}
\alias{make_0_to_1}
\title{Transform all the numeric variables in a dataframe so that their values are between 0 and 1 (impute with 0 or 1 if there is no variation).}
\usage{
make_0_to_1(x)
}
\arguments{
\item{x}{dataframe. Database with numeric variables to scale between 0 and 1.}
}
\value{
A dataframe where all numeric variables are between 0 and 1.
}
\description{
Transform all the numeric variables in a dataframe so that their values are between 0 and 1 (impute with 0 or 1 if there is no variation).
}
