% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cooc_net.R
\name{make_cooc_net}
\alias{make_cooc_net}
\title{Make a network based on cooccurrences of themes (use the document as unit of analysis).}
\usage{
make_cooc_net(x, dictionary)
}
\arguments{
\item{x}{data.frame. One numeric variable per theme, document ids as row names.}

\item{dictionary}{data.frame. Dictionary specifying the attributes of the themes, the variabme "name" identifying each theme.}
}
\value{
A network object
}
\description{
Make a network based on cooccurrences of themes (use the document as unit of analysis).
}
