% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dynamic_network.R
\name{make_dynamic_network}
\alias{make_dynamic_network}
\title{Create a dynamic network from periodic edgelists after filtering edges based on a weight criterion.}
\usage{
make_dynamic_network(x, period = "period", weight = "weight",
  threshold = 0.5, cores = 2)
}
\arguments{
\item{x}{dataframe. List of relationships for one group over several periods. It must have at least 4 columns: one for the period, "from" and "to" for the edges, and an attribute.}

\item{period}{character. Name of the variable indicating the period.}

\item{weight}{character. Name of the variable used to weight edges.}

\item{threshold}{numeric. Minimum value of the weighting attribute for the relationship to be maintained.}

\item{cores}{integer. Number of cores to be used for parallel computing.}
}
\value{
A dynamic network which can be rendered with render.d3movie
}
\description{
Create a dynamic network from periodic edgelists after filtering edges based on a weight criterion.
}
