% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_edgelist.R
\name{make_edgelist}
\alias{make_edgelist}
\title{Make an edgelist from a set of nodes attributes.}
\usage{
make_edgelist(x, nodes = NULL)
}
\arguments{
\item{x}{Tibble or dataframe. Table with one variable indicating the nodes.}

\item{nodes}{Character sting. Indicate the name of the variable used as node id.}
}
\value{
A tibble with, for each pair of nodes, the distance (for numeric variables) or cooccurrences (categorical variables).
}
\description{
Make an edgelist from a set of nodes attributes.
}
