% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_subgroups.R
\name{make_subgroups}
\alias{make_subgroups}
\title{Cluster observations using kmeans for either the specified number of subgroups or the number of unique observation.}
\usage{
make_subgroups(x, gpnbr)
}
\arguments{
\item{x}{dataframe. characteristics of the observation to group.}

\item{gpnbr}{integer. Number of subgroups.}
}
\value{
dataframe specifying group membership.
}
\description{
Cluster observations using kmeans for either the specified number of subgroups or the number of unique observation.
}
