% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unique.R
\name{make_unique}
\alias{make_unique}
\title{Remove duplicated entries. If there are still several instances of the same value in the id column, take the most frequent unique value, or the mean, or a random value of the duplicates to keep only one observation.}
\usage{
make_unique(x)
}
\arguments{
\item{x}{dataframe. Variables indicating the attributes of observations.}
}
\value{
Dataframe with only one instance of each id.
}
\description{
Remove duplicated entries. If there are still several instances of the same value in the id column, take the most frequent unique value, or the mean, or a random value of the duplicates to keep only one observation.
}
