% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_aggredge.R
\name{net_aggredge}
\alias{net_aggredge}
\title{Aggregate edges according to speciffications.}
\usage{
net_aggredge(edgelist, mode = "rolling", step = 1, range = 1)
}
\arguments{
\item{edgelist}{Tibble. Edgelist with four variables: year, from, to, weight}

\item{mode}{Character. Indicate whether edges should be aggregrated as "full", "cumulative", or "rolling".}

\item{step}{Numeric. Size of the increments in years.}

\item{range}{Numeric. In the "rolling" mode, this refers to the size of the window.}
}
\value{
An aggregated edgelist.
}
\description{
Aggregate edges according to speciffications.
}
