% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_compute_attributes.R
\name{net_compute_attributes}
\alias{net_compute_attributes}
\title{Compute network attributes for edges and nodes}
\usage{
net_compute_attributes(
  graph,
  directed = TRUE,
  weight_var = NULL,
  optimal_com = FALSE
)
}
\arguments{
\item{graph}{Tidygraph.}

\item{directed}{Logical. Whether the graph id directed.}

\item{weight_var}{Character. Name of the variable containing weights to be used}

\item{optimal_com}{Logical. Whether the optimal algorithm should be used to find communities.}
}
\value{
A tidygraph with network metrics for both edges and nodes.
}
\description{
Compute and append network attributes for both edges and nodes
}
\author{
Nicolas Mangin
}
