% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_compute_metrics.R
\name{net_compute_metrics}
\alias{net_compute_metrics}
\title{Add Jaccard coefficients to the edges of a tidygraph.}
\usage{
net_compute_metrics(graph, directed = TRUE, weight_var = NULL,
  optimal_com = FALSE)
}
\arguments{
\item{graph}{Tidygraph.}

\item{directed}{Logical. Whether the graph id directed.}

\item{weight_var}{Character. Name of the variable containing weights to be used}

\item{optimal_com}{Logical. Whether the optimal algorithm should be used to find communities.}
}
\value{
A tidygraph with network metrics for both edges and nodes.
}
\description{
Add Jaccard coefficients to the edges of a tidygraph.
}
