% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_compute_structure.R
\name{net_compute_structure}
\alias{net_compute_structure}
\title{Compute structural metrics for a network}
\usage{
net_compute_structure(graph, subgroup)
}
\arguments{
\item{graph}{Tidygraph.}

\item{subgroup}{Character. Name of the variable containing the community of the nodes in each graph.}
}
\value{
A tibble with network structure metrics for the graph.
}
\description{
Compute metrics describing the structure of a network
}
\author{
Nicolas Mangin
}
