% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_create_edgelist.R
\name{net_create_edgelist}
\alias{net_create_edgelist}
\title{Create an edgelist from attributes}
\usage{
net_create_edgelist(x, node_var = NULL)
}
\arguments{
\item{x}{Tibble or dataframe. Table with one variable indicating the nodes.}

\item{node_var}{Character sting. Indicate the name of the variable used as node id.}
}
\value{
A tibble with, for each pair of nodes, the distance (for numeric variables) or cooccurrences (categorical variables).
}
\description{
Create edges based on distances between attributes or based on common attributes.
}
\author{
Nicolas Mangin
}
