% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_create_graph.R
\name{net_create_graph}
\alias{net_create_graph}
\title{Create a graph from an edgelist}
\usage{
net_create_graph(edgelist, node_attributes = NULL)
}
\arguments{
\item{edgelist}{Tibble. Edge list; the two first variables have to be "from" and "to", the subsequent ones are treated as edge properties.}

\item{node_attributes}{Tibble. Properties of the nodes. The first variable has to be called "name" and contain the node id/label. Subsequent variables are treated as node properties.}
}
\value{
A tidygraph
}
\description{
Make an tidygraph from an edgelist and a set of node attributes
}
\author{
Nicolas Mangin
}
