% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_edge2graph.R
\name{net_edge2graph}
\alias{net_edge2graph}
\title{Make an tidygraph from an edgelist and a set of nodes attributes.}
\usage{
net_edge2graph(x, nodes_properties = NULL)
}
\arguments{
\item{x}{Tibble. Edge list; the two first variables have to be "from" and "to", the subsequent ones are treated as edge properties.}

\item{nodes_properties}{Tibble. Properties of the nodes. The first variable has to be called "name" and contain the node id/label. Subsequent variables are treated as node properties.}
}
\value{
A tidygraph
}
\description{
Make an tidygraph from an edgelist and a set of nodes attributes.
}
