% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_make_dynamic.R
\name{net_make_dynamic}
\alias{net_make_dynamic}
\title{Make a dynamic network}
\usage{
net_make_dynamic(
  graphlist,
  periodvar = "period",
  graphvar = "graph",
  node_weight = NULL,
  node_com = NULL,
  edge_weight = NULL,
  lscale = 0.05,
  vscale = 0.05,
  escale = 0.05,
  colors = c("red", "blue", "green", "orange", "purple", "darkgreen", "darkred")
)
}
\arguments{
\item{graphlist}{Tibble. One column must contain the "period" and another should contain the "graph"}

\item{periodvar}{Character. Name of the variable containing the period.}

\item{graphvar}{Character. Name of the variable containing the graph.}

\item{node_weight}{Character. Name of the variable containing the weights of the nodes in each graph.}

\item{node_com}{Character. Name of the variable containing the community of the nodes in each graph.}

\item{edge_weight}{Character. Name of the variable containing  the weights of the edges in each graph.}

\item{lscale}{Numeric. Scaling factor for labels.}

\item{vscale}{Numeric. Scaling factor for vertices.}

\item{escale}{Numeric. Scaling factor for edges.}

\item{colors}{Character vector. Colors of the communities.}
}
\value{
A D3 movie showing the evolution of the network.
}
\description{
Make a dynamic network from a list of graphs
}
\author{
Nicolas Mangin
}
