% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_metrics_overlap.R
\name{net_metrics_overlap}
\alias{net_metrics_overlap}
\title{Add Jaccard coefficients to the edges of a tidygraph.}
\usage{
net_metrics_overlap(graph)
}
\arguments{
\item{graph}{Tidygraph.}
}
\value{
A tidygraph with Jaccard coefficients.
}
\description{
Add Jaccard coefficients to the edges of a tidygraph.
}
