% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_net2egonet.R
\name{net_net2egonet}
\alias{net_net2egonet}
\title{Make an tidygraph from an edgelist and a set of nodes attributes.}
\usage{
net_net2egonet(graph, terms, separation = 2)
}
\arguments{
\item{graph}{Tidygraph.}

\item{terms}{Character. Regular expression to find the focal patterns.}

\item{separation}{Numeric. Maximum number of edges between focal and selected nodes.}
}
\value{
A tidygraph
}
\description{
Make an tidygraph from an edgelist and a set of nodes attributes.
}
