% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_semrel2net.R
\name{net_semrel2net}
\alias{net_semrel2net}
\title{Make an tidygraph from a list of semantic relationships as returned by text_txt2semrel.}
\usage{
net_semrel2net(semrel, basis = "lemma", keep_pos = c("NOUN", "PROPN",
  "ADJ", "VERB", "ADV"), multiplex = FALSE, min_count = 1)
}
\arguments{
\item{semrel}{Tibble. Output of the function text_txt2semrel}

\item{basis}{Character. Whether the network should be based on "word" or "lemma".}

\item{keep_pos}{Character vector. Parts of speech which should be kept to build the network.}

\item{multiplex}{Logical. Whether the kind of relationship should be kept.}

\item{min_count}{Numeric. Minimum number of occurrences for a relationship to be kept.}
}
\value{
A tidygraph of semantic relationships between words of lemmas
}
\description{
Make an tidygraph from a list of semantic relationships as returned by text_txt2semrel.
}
