% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_structure_metrics.R
\name{net_structure_metrics}
\alias{net_structure_metrics}
\title{Aggregate edges according to speciffications.}
\usage{
net_structure_metrics(graph, node_com)
}
\arguments{
\item{graph}{Tidygraph.}

\item{node_com}{Character. Name of the variable containing the community of the nodes in each graph.}
}
\value{
A tibble with network structure metrics for the graph.
}
\description{
Aggregate edges according to speciffications.
}
