% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_in_files.R
\name{replace_in_files}
\alias{replace_in_files}
\title{Replace specific strings of characters in a set of documents.}
\usage{
replace_in_files(path, pattern, replacement)
}
\arguments{
\item{path}{Character. Absolute path to the files.}

\item{pattern}{Character. Pattern to be replaced.}

\item{replacement}{Character. Pattern to be put instead.}
}
\value{
Replace in the same folder the documents where the string was found by the same documents where the string was replaced.
}
\description{
Replace specific strings of characters in a set of documents.
}
