% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_words.R
\name{replace_words}
\alias{replace_words}
\title{Replace words from a list by another word}
\usage{
replace_words(x, replacements)
}
\arguments{
\item{x}{character vector. Texts to clean.}

\item{replacements}{dataframe. Table with two variables: "pattern" for the string to replace and "replacement" for the string to put instead.}
}
\value{
A clean character vector.
}
\description{
Replace words from a list by another word
}
