% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cvar.R
\name{stat_cvar}
\alias{stat_cvar}
\title{Computes the coefficient of variation.}
\usage{
stat_cvar(x)
}
\arguments{
\item{x}{numeric vector. Set of values for which the coefficient of variation should be computed.}
}
\value{
A coefficient of variation for the numeric vector
}
\description{
Computes the coefficient of variation.
}
