% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_discretize.R
\name{stat_discretize}
\alias{stat_discretize}
\title{Make all the numeric variables of a dataframe discrete with the specified number of categories according to different algorithms (split based on quantiles or kmeans). The kmeans is chosen and if minimum differences are provided, differences inferior to the thresholds will be disregarded.}
\usage{
stat_discretize(x, ncat = 2, method = "km")
}
\arguments{
\item{x}{dataframe. database in which all numeric values have to be discretized.}

\item{ncat}{integer. The number of discrete values desired for each variable (ncat = 2 makes binary variables).}

\item{method}{character. Specify whether the variables should be discretized using quantiles ("qt") or kmeans ("km")}
}
\value{
A dataframe where all variables are discrete.
}
\description{
Make all the numeric variables of a dataframe discrete with the specified number of categories according to different algorithms (split based on quantiles or kmeans). The kmeans is chosen and if minimum differences are provided, differences inferior to the thresholds will be disregarded.
}
