% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_movavg.R
\name{stat_movavg}
\alias{stat_movavg}
\title{Computes the coefficient of variation.}
\usage{
stat_movavg(x, range)
}
\arguments{
\item{x}{numeric vector. Set of values to be transformed into moving averages.}

\item{range}{Integer. Number of (prior) periods to aggregate.}
}
\value{
A numeric vector of same size where values are smoothed.
}
\description{
Computes the coefficient of variation.
}
