% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tble_attri2edge.R
\name{tble_attri2edge}
\alias{tble_attri2edge}
\title{Make an edgelist from a set of nodes attributes.}
\usage{
tble_attri2edge(x, nodes = NULL)
}
\arguments{
\item{x}{Tibble or dataframe. Table with one variable indicating the nodes.}

\item{nodes}{Character sting. Indicate the name of the variable used as node id.}
}
\value{
A tibble with, for each pair of nodes, the distance (for numeric variables) or cooccurrences (categorical variables).
}
\description{
Make an edgelist from a set of nodes attributes.
}
