% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tble_get_cycle.R
\name{tble_get_cycle}
\alias{tble_get_cycle}
\title{Remove html tags, remove double spaces, convert to ASCII.}
\usage{
tble_get_cycle(x, range = 5, cutoff = c(-0.5, 0.5), minsep = 3)
}
\arguments{
\item{x}{Tibble. One column "group", "period" and one column "value".}

\item{range}{Integer. Number of years used to compute the slopes.}

\item{cutoff}{Numeric vector. Two values: minimum and maximum coefficient defining stability.}

\item{minsep}{Integer. Minimum number of periods of separation to justify a different phase.}
}
\value{
A tibble with cycle phases identified for each group in each period.
}
\description{
Remove html tags, remove double spaces, convert to ASCII.
}
