% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tble_interpolate.R
\name{tble_interpolate}
\alias{tble_interpolate}
\title{Replace missing values with interpolation (or extrapolation) over five years.}
\usage{
tble_interpolate(x)
}
\arguments{
\item{x}{Tibble. One column "year", and one column "value".}
}
\value{
A tibble with year and actual or interpolated values
}
\description{
Replace missing values with interpolation (or extrapolation) over five years.
}
