% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tble_remove_attributes.R
\name{tble_remove_attributes}
\alias{tble_remove_attributes}
\title{Remove attributes of all the variables.}
\usage{
tble_remove_attributes(x)
}
\arguments{
\item{x}{dataframe.}
}
\value{
Dataframe without attributes.
}
\description{
Remove attributes of all the variables.
}
