% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tble_unnest.R
\name{tble_unnest}
\alias{tble_unnest}
\title{Faster implementation of unnest. Applicable only when the nested column is a vector (i.e. not a table or dataframe).}
\source{
https://stackoverflow.com/questions/40420597/r-flatten-nested-data-table/40420690#40420690
}
\usage{
tble_unnest(x, col)
}
\arguments{
\item{x}{tibble. database with a column to unnest.}

\item{col}{character. Name of the column to unnest.}
}
\value{
A dataframe where all variables are discrete.
}
\description{
Faster implementation of unnest. Applicable only when the nested column is a vector (i.e. not a table or dataframe).
}
