% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tble_zero2one.R
\name{tble_zero2one}
\alias{tble_zero2one}
\title{Transform all the numeric variables in a dataframe so that their values are between 0 and 1 (impute with 0 or 1 if there is no variation).}
\usage{
tble_zero2one(x)
}
\arguments{
\item{x}{dataframe. Database with numeric variables to scale between 0 and 1.}
}
\value{
A dataframe where all numeric variables are between 0 and 1.
}
\description{
Transform all the numeric variables in a dataframe so that their values are between 0 and 1 (impute with 0 or 1 if there is no variation).
}
