% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_anno2bow.R
\name{text_anno2bow}
\alias{text_anno2bow}
\title{Transform the documents stored in a dataframe in a bag of words.}
\usage{
text_anno2bow(x, base = "lemma", keep = c("NOUN", "PROPN", "ADJ",
  "VERB", "ADV"))
}
\arguments{
\item{x}{tibble. Output of the function text_annotate.}

\item{base}{character string. whether a bag of "word"s or a bag of "lemma"s should be produced.}

\item{keep}{character vector. Which parts of speech should be keept when building the bag of words.}
}
\value{
A tibble with the document, term, and count.
}
\description{
Transform the documents stored in a dataframe in a bag of words.
}
