% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_annotate.R
\name{text_annotate}
\alias{text_annotate}
\title{Apply an udpipe model for part-of-speech tagging.}
\usage{
text_annotate(x, model)
}
\arguments{
\item{x}{character. String to be tagged.}

\item{model}{character. path from the working directory to the downloaded udpipe model.}
}
\value{
A dataframe where each token is a lemmatized and tagged observation. Tags for parts of speech are: "NOUN", "PROPN", "ADJ", "VERB", "ADV", "AUX", "PRON", "DET", "PART", "ADP", "CCONJ", "SCONJ", "NUM", "SYM", "PUNCT", "X"
}
\description{
Apply an udpipe model for part-of-speech tagging.
}
