% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_apply_dictionaries.R
\name{text_apply_dictionaries}
\alias{text_apply_dictionaries}
\title{Count the occurrences of words associated with a dictionary.}
\usage{
text_apply_dictionaries(bow, dictionary, score = "total")
}
\arguments{
\item{bow}{Tibble. Bag of words for one document with at least two variables: "term" and "count".}

\item{dictionary}{Tibble. Categorizations of words with at least three variables: "term", "category", and "value".}

\item{score}{Character. Whether the function should return a "total" or a "percentage".}
}
\value{
A tibble with documents as rows, categories as variables, ad the required summary in cells
}
\description{
Count the occurrences of words associated with a dictionary.
}
