% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_apply_stm.R
\name{text_apply_stm}
\alias{text_apply_stm}
\title{Apply the topic mode to a new corpus of documents.}
\usage{
text_apply_stm(bow, stm, min_gamma, max_topic)
}
\arguments{
\item{bow}{Tibble. Bag of words as returned by the function text_semrel2bow.}

\item{stm}{stm. Structural topic model as returned by the function stm.}

\item{min_gamma}{Numeric. Minumum strength of the relationship between topic and socument.}

\item{max_topic}{Integer. Maximum number of topics allowed per document.}
}
\value{
A tibble linking topics and documents for a new corpus.
}
\description{
Apply the topic mode to a new corpus of documents.
}
