% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_bow2dtm.R
\name{text_bow2dtm}
\alias{text_bow2dtm}
\title{Transform the bags of words into a document to term matrix/}
\usage{
text_bow2dtm(x, document = "document", min_term_count = 0,
  min_term_src_count = 0, nbterm = 1000, docvar = NULL)
}
\arguments{
\item{x}{tibble. Output of the function text_bow_metrics.}

\item{document}{character. name of the variable identifying documents.}

\item{min_term_count}{Integer. Remove terms appearing less than this number of times.}

\item{min_term_src_count}{Integer, Remove terms appearing in less than this number of documents.}

\item{nbterm}{Integer. Select this number of terms based on tf idf.}

\item{docvar}{Additional information about documents to be appended to the docvar of the dtm.}
}
\value{
A document to term matrix.
}
\description{
Transform the bags of words into a document to term matrix/
}
