% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_bow_metrics.R
\name{text_bow_metrics}
\alias{text_bow_metrics}
\title{Compute basic metrics for a bag of words.}
\usage{
text_bow_metrics(bow, document = "document", term = "term",
  count = "count")
}
\arguments{
\item{bow}{tibble. Bag of words produced with text_semrel2bow}

\item{document}{character string. Name of the variable indicating the document id.}

\item{term}{character string. Name of the variable indicating the term.}

\item{count}{character string. Name of the variable indicating the frequence.}
}
\value{
A tibble 9 variables: document, term, count, doc_word_count (number of words in the document), term_count (frequency of the term), term_document_count (number of documents in which the term appears), tf, idf, tf_idf.
}
\description{
Compute basic metrics for a bag of words.
}
