% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_clean.R
\name{text_clean}
\alias{text_clean}
\title{Remove html tags, remove double spaces, convert to ASCII.}
\usage{
text_clean(text)
}
\arguments{
\item{text}{character. String to be cleaned.}
}
\value{
A clean string.
}
\description{
Remove html tags, remove double spaces, convert to ASCII.
}
