% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_compare.R
\name{text_compare}
\alias{text_compare}
\title{Transform the documents stored in a dataframe in a bag of words.}
\usage{
text_compare(x, y)
}
\arguments{
\item{x}{character string.}

\item{y}{character string.}
}
\value{
A tibble indicating elements of comparison between x and y contents.
}
\description{
Transform the documents stored in a dataframe in a bag of words.
}
