% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_eval_read.R
\name{text_eval_read}
\alias{text_eval_read}
\title{Assess the readability of a text using various indicators.}
\usage{
text_eval_read(text)
}
\arguments{
\item{text}{character. Text.}
}
\value{
Various readability scores as one observation in a tibble.
}
\description{
Assess the readability of a text using various indicators.
}
