% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_eval_stm.R
\name{text_eval_stm}
\alias{text_eval_stm}
\title{Return various metrics to evaluate topic model performance for different topic numbers.}
\usage{
text_eval_stm(dtm = NA, topic_nbr = c(0), prevalence = NA,
  init.type = "Spectral", seed_nbr = 1, keywords = NA,
  topic_per_doc = 5, terms_per_topic = 5)
}
\arguments{
\item{dtm}{dtm object.}

\item{topic_nbr}{Numeric vector}

\item{prevalence}{Formula}

\item{init.type}{Character.}

\item{seed_nbr}{Numeric vector}

\item{keywords}{Tibble.}

\item{topic_per_doc}{Integer.}

\item{terms_per_topic}{Integer.}
}
\value{
A tibble with various metrics of tm quality for each possible combination of the topic number and seed
}
\description{
Return various metrics to evaluate topic model performance for different topic numbers.
}
