% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extract_around.R
\name{text_extract_around}
\alias{text_extract_around}
\title{Extract windows of words around a focal worj}
\usage{
text_extract_around(text, term, window = c(3, 3))
}
\arguments{
\item{text}{character. Text.}

\item{term}{character. regex pattern around which the adjacent pattern has to be detected within the window.}

\item{window}{numeric vector. Two values: one indicating the beginning of the window from the focal pattern, and one indicating the end.}
}
\value{
A list of windows around each occurrences of the focal term.
}
\description{
Extract windows of words around a focal worj
}
