% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extract_window.R
\name{text_extract_window}
\alias{text_extract_window}
\title{Extract windows of words around a focal word}
\usage{
text_extract_window(text, term, window = c(3, 3))
}
\arguments{
\item{text}{Character. String from which the indows should be extracted.}

\item{term}{Character. regex pattern around which the adjacent pattern has to be detected within the window.}

\item{window}{Numeric vector. Two values: one indicating the beginning of the window from the focal pattern, and one indicating the end.}
}
\value{
A list of windows around each occurrences of the focal term.
}
\description{
Extract windows of words around a focal word
}
