% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_keep_letters.R
\name{text_keep_letters}
\alias{text_keep_letters}
\title{Remove all characthers but letters.}
\usage{
text_keep_letters(text)
}
\arguments{
\item{text}{character. String to be cleaned.}
}
\value{
A clean string.
}
\description{
Remove all characthers but letters.
}
