% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_model_topics.R
\name{text_model_topics}
\alias{text_model_topics}
\title{Apply the topic mode to a new corpus of documents.}
\usage{
text_model_topics(dtm, topic_nbr = 0, prevalence = NULL,
  content = NULL, init.type = "Spectral", seed = 1234,
  keywords = NULL, terms_per_topic = 5, baselab = "prob",
  frexweight = 0.6, min_gamma = 0.1, min_common = 1)
}
\arguments{
\item{dtm}{dtm.}

\item{topic_nbr}{Numeric.}

\item{prevalence}{Formula}

\item{content}{Formula}

\item{init.type}{Character.}

\item{seed}{Numeric}

\item{keywords}{Tibble.}

\item{terms_per_topic}{Integer.}

\item{baselab}{Character. Whether labels should be based on "prob", "frex", "lift", or "score".}

\item{frexweight}{Numeric. The weight on discriminance in term selection.}

\item{min_gamma}{Numeric. Minimum strength of the relationship between document and topic.}

\item{min_common}{Integer. Minimum number of keywords in the label.}
}
\value{
A list with the "topic_model", the "topic_label", and the "document_topic".
}
\description{
Apply the topic mode to a new corpus of documents.
}
