% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_overlap.R
\name{text_overlap}
\alias{text_overlap}
\title{Transform the documents stored in a dataframe in a bag of words.}
\usage{
text_overlap(src, x, y)
}
\arguments{
\item{src}{character string. id of the source text to which the list of strings is compared.}

\item{x}{character string. source text to which the list of target strings is compared.}

\item{y}{list od strings. list of target strings to which the source string is compared.}
}
\value{
A tibble indicating the overlaps between strings.
}
\description{
Transform the documents stored in a dataframe in a bag of words.
}
