% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_parse.R
\name{text_parse}
\alias{text_parse}
\title{Apply an udpipe model for part-of-speech tagging.}
\usage{
text_parse(x, model)
}
\arguments{
\item{x}{character. String to be parsed.}

\item{model}{character. path from the working directory to the downloaded udpipe model.}
}
\value{
A dataframe where each relationship between word is an observation
}
\description{
Apply an udpipe model for part-of-speech tagging.
}
