% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_parsedep.R
\name{text_parsedep}
\alias{text_parsedep}
\title{Apply an udpipe model for part-of-speech tagging.}
\usage{
text_parsedep(document, text, model)
}
\arguments{
\item{document}{character. ID of the document.}

\item{text}{character. String to be parsed.}

\item{model}{character. path from the working directory to the downloaded udpipe model.}
}
\value{
A dataframe where each relationship between word is an observation
}
\description{
Apply an udpipe model for part-of-speech tagging.
}
