% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_recompose.R
\name{text_recompose}
\alias{text_recompose}
\title{Identify linebreaks which do not seem to match an end of paragraph and recompose the paragraph.}
\usage{
text_recompose(document)
}
\arguments{
\item{document}{Character vector. A vector of lines of text to be recomposed into paragraphs.}
}
\value{
A character vector in which paragraphs (ending with "." and with an empty line after) are recomposed.
}
\description{
Identify linebreaks which do not seem to match an end of paragraph and recompose the paragraph.
}
