% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_replace_words.R
\name{text_replace_words}
\alias{text_replace_words}
\title{Identify linebreaks which do not seem to match an end of paragraph and recompose the paragraph.}
\usage{
text_replace_words(text, dictionary)
}
\arguments{
\item{text}{Character. A string or vector of strings in which patterns have to be replaced.}

\item{dictionary}{Tibble. A tibble with two variables: "pattern" and "replacement".}
}
\value{
The initial string or vector in which replacements have been made.
}
\description{
Identify linebreaks which do not seem to match an end of paragraph and recompose the paragraph.
}
