% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_score_dictionaries.R
\name{text_score_dictionaries}
\alias{text_score_dictionaries}
\title{Count the occurrences of words associated with a dictionary.}
\usage{
text_score_dictionaries(x, type = "vector", dictionary,
  category = NULL, pattern = NULL)
}
\arguments{
\item{x}{String or character vector. Content to be scored.}

\item{type}{Character. Whether x is a "string" (in which case it will be split) or a "vector"}

\item{dictionary}{Tibble. Categorizations of words.}

\item{category}{Character. Name of the grouping variable in the dictionary.}

\item{pattern}{Character. Name of the pattern variable in the dictionary.}
}
\value{
A tibble with the count of words, for each category and in total, forht the document.
}
\description{
Count the occurrences of words associated with a dictionary.
}
