% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_score_sentiment.R
\name{text_score_sentiment}
\alias{text_score_sentiment}
\title{Count the occurrences of words associated with a dictionary.}
\usage{
text_score_sentiment(x, dictionary, type = "vector", level = "cluster")
}
\arguments{
\item{x}{String or character vector. Conte to be scored.}

\item{dictionary}{Tibble. Categorizations of words.}

\item{type}{Character. Whether x is a "string" (in which case it will be split) or a "vector"}

\item{level}{Character. Name of the grouping variable in the dictionary.}
}
\value{
A tibble with the count of words, for each category and in total, forht the document.
}
\description{
Count the occurrences of words associated with a dictionary.
}
