% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_simplify.R
\name{text_simplify}
\alias{text_simplify}
\title{Simplify text to facilitate comparisons of strings}
\usage{
text_simplify(text)
}
\arguments{
\item{text}{character. String to be simplified}
}
\value{
A siplified string.
}
\description{
Simplify text to facilitate comparisons of strings
}
