% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_test_regex.R
\name{text_test_regex}
\alias{text_test_regex}
\title{Gadget to test and ajust regular expressions in a dictionary.}
\usage{
text_test_regex(dictionary, corpus)
}
\arguments{
\item{dictionary}{Tibble. Dictionary with at least a column "pattern".}

\item{corpus}{List. List of texts.}
}
\value{
A new dictionary
}
\description{
Gadget to test and ajust regular expressions in a dictionary.
}
