% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_txt2semrel.R
\name{text_txt2semrel}
\alias{text_txt2semrel}
\title{Apply a udpipe model for part-of-speech tagging and dependency parsing.}
\usage{
text_txt2semrel(text, model)
}
\arguments{
\item{text}{character. String to be parsed.}

\item{model}{character. path from the working directory to the downloaded udpipe model.}
}
\value{
A dataframe where each observation is a relationship between word from the initial text.
}
\description{
Apply a udpipe model for part-of-speech tagging and dependency parsing.
}
