% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_unbreak.R
\name{text_unbreak}
\alias{text_unbreak}
\title{Transform the documents stored in a dataframe in a bag of words.}
\usage{
text_unbreak(x)
}
\arguments{
\item{x}{character vector in which paragraphs are over multiple lines.}
}
\value{
A character vector in which paragraphs (ending with . and with an empty line after) are recomposed.
}
\description{
Transform the documents stored in a dataframe in a bag of words.
}
