% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf_discretize.R
\name{transf_discretize}
\alias{transf_discretize}
\title{Transform a continuous variable into a discrete one}
\usage{
transf_discretize(x, method = "quantiles", nval = 2)
}
\arguments{
\item{x}{Numeric. Vector of values to be transformed into a discrete ordinal variable.}

\item{method}{character. Specify whether the variables should be discretized using quantiles ("quantiles") or kmeans ("kmeans")}

\item{nval}{Integer. Number of discrete values desired for each variable (groups = 2 makes binary variables).}
}
\value{
A dataframe where all variables are discrete.
}
\description{
Transform a continuous variable into a discrete one.
}
\author{
Nicolas Mangin
}
