% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf_impute_missing.R
\name{transf_impute_missing}
\alias{transf_impute_missing}
\title{Impute missing values}
\usage{
transf_impute_missing(x)
}
\arguments{
\item{x}{Numeric or character vector with missing values.}
}
\value{
A vector of same class without missing values.
}
\description{
Impute numeric values with the mean and categorical values with the most frequent category.
}
\author{
Nicolas Mangin
}
