% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf_interpolate_missing.R
\name{transf_interpolate_missing}
\alias{transf_interpolate_missing}
\title{Interpolate or extrapolate missing values}
\usage{
transf_interpolate_missing(x, range)
}
\arguments{
\item{x}{Numeric. Vector of chronological values with missing data.}

\item{range}{Integer. Number of periods before and after used to interpolate with the focal value.}
}
\value{
A tibble with year and actual or interpolated values
}
\description{
Replace missing values with interpolation (or extrapolation) over five years.
}
\author{
Nicolas Mangin
}
