% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf_moving_aggregates.R
\name{transf_moving_aggregates}
\alias{transf_moving_aggregates}
\title{Compute moving aggregates}
\usage{
transf_moving_aggregates(x, range, fun = NULL)
}
\arguments{
\item{x}{Numeric. Vector of chronological values to be transformed into moving averages.}

\item{range}{Integer. Number of periods before and after to average with the focal value.}

\item{fun}{Function. Function to apply on the series (e.g. sum, mean, sd or any function taking a numeric vector as input and returns a numeric value)}
}
\value{
A numeric vector of same size where values are smoothed.
}
\description{
Compute moving averages or a chronological numeric vector
}
\author{
Nicolas Mangin
}
