% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf_moving_averages.R
\name{transf_moving_averages}
\alias{transf_moving_averages}
\title{Compute moving average values}
\usage{
transf_moving_averages(x, range)
}
\arguments{
\item{x}{Numeric. Vector of chronological values to be transformed into moving averages.}

\item{range}{Integer. Number of periods before and after to average with the focal value.}
}
\value{
A numeric vector of same size where values are smoothed.
}
\description{
Compute moving averages or a chronological numeric vector
}
\author{
Nicolas Mangin
}
