% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf_zero_to_one.R
\name{transf_zero_to_one}
\alias{transf_zero_to_one}
\title{Tranform from 0 to 1}
\usage{
transf_zero_to_one(x)
}
\arguments{
\item{x}{Numeric. Values to be transformed into 0 to 1 values}
}
\value{
A numeric variables going from 0 and 1.
}
\description{
Transform a numeric variable so that its values are between 0 and 1 (impute with 1 if no variation)
}
\author{
Nicolas Mangin
}
