% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim.R
\name{aim}
\alias{aim}
\title{Set Makefile target}
\usage{
aim(target = NULL)
}
\value{
No return value. Called for side effects.
}
\description{
\code{aim} looks for an existing \code{Makefile}, reads its content, and
offers a list of discovered \code{Makefile} "rules" or "targets" (build
scripts, in our case) in an interactive way.
}
\examples{
\dontrun{
aim()
}
}
\seealso{
Other functions from \code{buildr} trinity: 
\code{\link{build}()},
\code{\link{init}()}
}
\author{
Jan Netik
}
\concept{functions from \code{buildr} trinity}
\keyword{file}
\keyword{misc}
\keyword{utilities}
