% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim.R
\name{aim}
\alias{aim}
\title{Set Makefile Target}
\usage{
aim(target = NULL)
}
\arguments{
\item{target}{\emph{Character}. The name of the Makefile target to set.}
}
\value{
No return value. Called for side effects.
}
\description{
\code{aim()} looks for an existing \code{Makefile}, reads its content, and offers a
list of discovered \code{Makefile} targets (denoting build scripts, in our case),
all in an interactive way. When the session is not interactive, or you know
the name of the desired target, you can declare it directly in the \code{target}
argument.
}
\examples{
\dontrun{
# We have several build scripts in our project root
# and we want to select script called "build_all.R":

aim(target = "all") # note that "build_" is stripped out by default
}
}
\seealso{
Other functions from buildr trinity: 
\code{\link{build}()},
\code{\link{init}()}
}
\author{
Jan Netik
}
\concept{functions from buildr trinity}
\keyword{file}
\keyword{misc}
\keyword{utilities}
