% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Run Selected Build Script}
\usage{
build()
}
\value{
Character string echoing the terminal.
}
\description{
Sbalbalbaab.
}
\examples{
\dontrun{
build()
}
}
\seealso{
Other functions from \code{buildr} trinity: 
\code{\link{aim}()},
\code{\link{init}()}
}
\author{
Jan Netik
}
\concept{functions from \code{buildr} trinity}
\keyword{file}
\keyword{misc}
\keyword{utilities}
