% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Build}
\usage{
build()
}
\value{
Character string echoing the terminal.
}
\description{
Sbalbalbaab.
}
\examples{
\dontrun{
build()
}
}
\seealso{
Other functions from \code{buildr} trinity: 
\code{\link{aim}()},
\code{\link{init}()}
}
\author{
Jan Netik \cr

Department of Psychology, \cr Faculty of Arts, \cr
Charles University, \cr Czech Republic \cr

\email{netikja@gmail.com} \cr
}
\concept{functions from \code{buildr} trinity}
\keyword{file}
\keyword{misc}
\keyword{utilities}
