% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Run Selected Build Script}
\usage{
build()
}
\value{
No return value. Called for side effects.
}
\description{
\code{build()} is the final function in the workflow, as it instructs 'RStudio'
Build pane to take the first rule in the \code{Makefile} (set previously with
\code{aim()}) and runs the respective recipe.
}
\details{
The 'Rstudio' Build pane is not allways visible and set to take \code{Makefiles}.
However, the \code{build()} ensures that everything is set properly and if not, it
offers you to automatically edit necessary settings automatically for you.
Note that this action forces 'RStudio' user interface (UI) to reload and you
have to call \code{build()} again afterwards.
}
\examples{
\dontrun{
build()
}
}
\seealso{
Other functions from buildr trinity: 
\code{\link{aim}()},
\code{\link{init}()}
}
\author{
Jan Netik
}
\concept{functions from buildr trinity}
\keyword{file}
\keyword{misc}
\keyword{utilities}
