% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendly.R
\name{build_binaries}
\alias{build_binaries}
\title{Build a set of binaries, waiting until complete}
\usage{
build_binaries(filenames, config, dest = tempfile())
}
\arguments{
\item{filenames}{Vector of filenames of the source packages}

\item{config}{Configuation options; a list with options "host"
(required), port, poll and timeout (all optional and passed to
\code{\link{buildr_client}} and \code{buildr_client$wait}).
Alternatively, if \code{config=FALSE}, this will build binaries
using your system directly rather than using a build server.}

\item{dest}{A directory to place generated files.}
}
\value{
A character vector of filenames
}
\description{
Build a set of binaries, waiting until complete
}

