% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildr-package.R
\docType{package}
\name{buildr-package}
\alias{buildr}
\alias{buildr-package}
\title{buildr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig/buildr}
  \item \url{http://docs.jimbrig.com/buildr/}
  \item Report bugs at \url{https://github.com/jimbrig/buildr/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

}
\keyword{internal}
