% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildr-package.R
\docType{package}
\name{buildr-package}
\alias{buildr}
\alias{buildr-package}
\title{buildr: Organize & Run Build Scripts Comfortably}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Working with reproducible reports or any other similar projects often require to run the script that builds the output file in a specified way. 'buildr' can help you organize, modify and comfortably run those scripts. The package provides a set of functions that interactively guides you through the process and that are available as 'RStudio' Addin, meaning you can set up the keyboard shortcuts, enabling you to choose and run the desired build script with one keystroke anywhere anytime.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://netique.github.io/buildr/}
  \item Report bugs at \url{https://github.com/netique/buildr/issues/}
}

}
\author{
\strong{Maintainer}: Jan Netik \email{netikja@gmail.com} (\href{https://orcid.org/0000-0002-3888-3203}{ORCID})

}
\keyword{internal}
