% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildr.R
\name{buildr}
\alias{buildr}
\title{Runs \code{build.R} script}
\usage{
buildr()
}
\value{
No return value.
}
\description{
Sources the \code{build.R} (case \strong{insensitive}) present in current working directory. Print \code{buildr()} in the console or click on 'Buildr' in addins menu in 'RStudio'.
}
\examples{
\dontrun{
buildr()
}
}
\author{
Jan Netik \cr

Department of Psychology, \cr
Faculty of Arts, \cr
Charles University, \cr
Czech Republic \cr

\email{netikja@gmail.com} \cr
}
