% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in data file in csv format and convert to tibble}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character vector matching the name of the file to be imported}
}
\value{
This function returns a dataset in tibble format. As a side effect, this function
   also prints the tibble.
}
\description{
This is a function that imports a comma-separated values file from the working directory
via \code{read_csv} from \code{readr}. After import, the data object is converted
to tibble format via \code{tbl_df} from \code{dplyr}.
}
\note{
If the file does not exist in the working directory, the function will return an error.
}
\examples{
\dontrun{fars_read("accident_2013.csv")}

}
