% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Return a list of tibbles from a list of specified years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of numeric values matching the corresponding csv files to import}
}
\value{
This function has a side effect that returns a list of tibbles with two columns: year and MONTH
}
\description{
This function takes a list of numeric values and returns a list of tibbles
containing the MONTH and year columns only. The list of numeric values are
translated to filenames via \code{make_filename} and the filenames are used
to read in the corresponding csv files via \code{fars_read}. The tibbles
returned by \code{fars_read} are reduced to the MONTH and year columns only.
}
\note{
If the argument of the years is not a list or if the list of tibbles cannot be executed,
   the function will return an error, specifying which year could not be returned.
}
\examples{
\dontrun{years <- list(2013, 2014, 2015)
fars_read_years(years = years)}

}
