% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize a list of tibbles by observation counts}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of numeric values matching the corresponding csv files to import and summarize}
}
\value{
This function returns a summary data frame in wide format as a side effect.
}
\description{
This function summarizes the list of tibbles returned by \link{\code{fars_read_years}}.
The function aggregates the list of tibbles into a single tibble, creates groups by year
and MONTH, counts the number of observations into a new column "n", and returns the summary
tibble in wide format.
}
\examples{
\dontrun{years <- list(2013, 2014, 2015)
fars_summarize_years(years = years)}

}
