% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize}
\usage{
init(prefix = "build", path = ".", sep = "_", command_args = "")
}
\value{
Character string echoing the terminal.
}
\description{
Sbalbalbaab.
}
\examples{
\dontrun{
init()
}
}
\seealso{
Other functions from \code{buildr} trinity: 
\code{\link{aim}()},
\code{\link{build}()}
}
\author{
Jan Netik
}
\concept{functions from \code{buildr} trinity}
\keyword{file}
\keyword{misc}
\keyword{utilities}
