% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize}
\usage{
init(prefix = "build", path = ".", sep = "_", command_args = "")
}
\value{
Character string echoing the terminal.
}
\description{
Sbalbalbaab.
}
\examples{
\dontrun{
init()
}
}
\author{
Jan Netik \cr

Department of Psychology, \cr Faculty of Arts, \cr
Charles University, \cr Czech Republic \cr

\email{netikja@gmail.com} \cr
}
