% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a filename given a provided year(s)}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric value of any length}
}
\value{
This function produces a side effect by printing the generated filename as a character string
}
\description{
This function takes a single value, vector or list representing a four integer year. The function prints
a character string with the following format: "accident_[year].csv.bz2" for each value provided.
}
\examples{
\dontrun{make_filename(2018)}

}
