% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_compute_overlap.R
\name{net_compute_overlap}
\alias{net_compute_overlap}
\title{Compute overlap metrics}
\usage{
net_compute_overlap(graph)
}
\arguments{
\item{graph}{Tidygraph.}
}
\value{
A tidygraph with Jaccard coefficients.
}
\description{
Add Jaccard coefficients and other overlap metrics to the edges of a tidygraph.
}
\author{
Nicolas Mangin
}
