% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_bow2dtm.R
\name{text_bow2dtm}
\alias{text_bow2dtm}
\title{Transform the bags of words into a document to term matrix/}
\usage{
text_bow2dtm(bow, min_term = 0, max_term = Inf, min_doc = 0,
  max_doc = Inf, nbterm = 1000, docvar = NULL)
}
\arguments{
\item{bow}{Tibble. Output of the function text_bow_metrics. Document ids must be in a variable called "document".}

\item{min_term}{Integer. Remove terms appearing less than this number of times.}

\item{max_term}{Integer. Remove terms appearing more than this number of times.}

\item{min_doc}{Integer, Remove terms appearing in less than this number of documents.}

\item{max_doc}{Integer, Remove terms appearing in more than this number of documents.}

\item{nbterm}{Integer. Select this number of terms based on tf idf.}

\item{docvar}{Additional information about documents to be appended to the docvar of the dtm. Document ids must be in a variable called "document".}
}
\value{
A document to term matrix.
}
\description{
Transform the bags of words into a document to term matrix/
}
