% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_bow_metrics.R
\name{text_bow_metrics}
\alias{text_bow_metrics}
\title{Compute basic metrics for a bag of words.}
\usage{
text_bow_metrics(x, source = "source", term = "term",
  count = "count")
}
\arguments{
\item{x}{tibble. Bag of words produced with text_anno2bow}

\item{source}{character string. Name of the variable indicating the document id.}

\item{term}{character string. Name of the variable indicating the term.}

\item{count}{character string. Name of the variable indicating the frequence.}
}
\value{
A tibble 9 variables: source, term, count, src_word_count (number of words in the source), term_count (frequency of the term), term_source_count (number of sources in which the term appears), tf, idf, tf_idf.
}
\description{
Compute basic metrics for a bag of words.
}
