% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_clean.R
\name{text_clean}
\alias{text_clean}
\title{Simplify and clean a text.}
\usage{
text_clean(x)
}
\arguments{
\item{x}{character. String to be cleaned.}
}
\value{
A clean string.
}
\description{
Simplify and clean a text.
}
