% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_eval_stm.R
\name{text_eval_stm}
\alias{text_eval_stm}
\title{Return various metrics to evaluate topic model performance for different topic numbers.}
\usage{
text_eval_stm(dtm = NA, topic_nbr = 0, prevalence = NULL,
  content = NULL, init.type = "Spectral", seed = 1,
  keywords = NULL, min_beta = 0.01, min_gamma = 0.1,
  min_common = 1)
}
\arguments{
\item{dtm}{dtm.}

\item{topic_nbr}{Numeric.}

\item{prevalence}{Formula}

\item{content}{Formula}

\item{init.type}{Character.}

\item{seed}{Numeric}

\item{keywords}{Tibble.}

\item{min_beta}{Numeric. Minimum strength of the relationship between term and topic.}

\item{min_gamma}{Numeric. Minimum strength of the relationship between document and topic.}

\item{min_common}{Integer. Maximum number of topics allowed per document.}
}
\value{
A list with the "topic_model", the "topic_label", and the "document_topic".

A tibble with various metrics of tm quality for each possible combination of the topic number and seed
}
\description{
Return various metrics to evaluate topic model performance for different topic numbers.
}
