% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_eval_stm.R
\name{text_eval_stm}
\alias{text_eval_stm}
\title{Return various metrics to evaluate topic model performance for different topic numbers.}
\usage{
text_eval_stm(dtm = NA, topic_nbr = 0, prevalence = NULL,
  content = NULL, init.type = "Spectral", seed = 1,
  keywords = NULL, terms_per_topic = 5)
}
\arguments{
\item{dtm}{dtm.}

\item{topic_nbr}{Numeric.}

\item{prevalence}{Formula}

\item{content}{Formula}

\item{init.type}{Character.}

\item{seed}{Numeric}

\item{keywords}{Tibble.}

\item{terms_per_topic}{Integer.}
}
\value{
A tibble with various metrics of tm quality for each possible combination of the topic number and seed
}
\description{
Return various metrics to evaluate topic model performance for different topic numbers.
}
