% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_model_topics.R
\name{text_model_topics}
\alias{text_model_topics}
\title{Apply the topic mode to a new corpus of documents.}
\usage{
text_model_topics(
  dtm,
  topic_nbr = 0,
  prevalence = NULL,
  content = NULL,
  init.type = "Spectral",
  seed = 1234,
  iterations = 50,
  tolerance = 0,
  keywords = NULL,
  baselab = "prob",
  frexweight = 0.6,
  min_gamma = 0.1,
  min_beta = 0.01,
  min_common = 1,
  verbose = TRUE
)
}
\arguments{
\item{dtm}{dtm.}

\item{topic_nbr}{Numeric.}

\item{prevalence}{Formula}

\item{content}{Formula}

\item{init.type}{Character.}

\item{seed}{Numeric.}

\item{iterations}{Integer.}

\item{tolerance}{Numeric.}

\item{keywords}{Tibble.}

\item{baselab}{Character. Whether labels should be based on "prob", "frex", "lift", or "score".}

\item{frexweight}{Numeric. The weight on discriminance in term selection.}

\item{min_gamma}{Numeric. Minimum strength of the relationship between document and topic.}

\item{min_beta}{Numeric. Minimum strength of the relationship between term and topic.}

\item{min_common}{Integer. Maximum number of topics allowed per document.}

\item{verbose}{Logical.}
}
\value{
A list with the "topic_model", the "topic_term", the "topic_label", and the "topic_document".
}
\description{
Apply the topic mode to a new corpus of documents.
}
