% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_score_dictionaries.R
\name{text_score_dictionaries}
\alias{text_score_dictionaries}
\title{Count the occurrences of words associated with a dictionary.}
\usage{
text_score_dictionaries(x, dictionary, category = NULL,
  type = "vector")
}
\arguments{
\item{x}{String or character vector. Conte to be scored.}

\item{dictionary}{Tibble. Categorizations of words.}

\item{category}{Character. Name of the grouping variable in the dictionary.}

\item{type}{Character. Whether x is a "string" (in which case it will be split) or a "vector"}
}
\value{
A tibble with the count of words, for each category and in total, forht the document.
}
\description{
Count the occurrences of words associated with a dictionary.
}
