% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_semrel2bow.R
\name{text_semrel2bow}
\alias{text_semrel2bow}
\title{Extract bags of words from semating relationships.}
\usage{
text_semrel2bow(semrel, basis = "lemma", keep_pos = c("NOUN", "PROPN",
  "ADJ", "VERB", "ADV"))
}
\arguments{
\item{semrel}{Tibble. Semantic for one document relationships as returned by text_txt2semrel.}

\item{basis}{Character. Whether a bag of "word", a bag of "lemma", or a bag of "stem" should be returned.}

\item{keep_pos}{Character vector. Parts of speech which should be kept to build the bags of words}
}
\value{
A tibble simplified into a bag of words.
}
\description{
Extract bags of words from semating relationships.
}
